+++ Shared/Managers/LaunchManager.swift
@@ -54,7 +54,7 @@ class LaunchManager {
             Task {
                 await UserNotificationManager.shared.silentRegister(withApplication: UIApplication.shared)
             }
-            // Pre - load contacts 
+            // Pre-load contacts
             _ = ContactsManger.shared
         }
 #endif
